requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==24.1.0
    # via quart
aiokafka==0.12.0
    # via ugc-1-sprint-8 (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
async-timeout==5.0.1
    # via
    #   aiokafka
    #   redis
blinker==1.9.0
    # via
    #   flask
    #   quart
click==8.2.1
    # via
    #   flask
    #   quart
clickhouse-driver==0.2.9
    # via ugc-1-sprint-8 (pyproject.toml)
exceptiongroup==1.3.0
    # via
    #   hypercorn
    #   pytest
    #   taskgroup
flask==3.1.1
    # via quart
googleapis-common-protos==1.59.1
    # via opentelemetry-exporter-jaeger-proto-grpc
grpcio==1.74.0
    # via opentelemetry-exporter-jaeger-proto-grpc
h11==0.16.0
    # via
    #   hypercorn
    #   wsproto
h2==4.2.0
    # via hypercorn
hpack==4.1.0
    # via h2
hypercorn==0.17.3
    # via quart
hyperframe==6.1.0
    # via h2
importlib-metadata==8.7.0
    # via opentelemetry-api
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via
    #   flask
    #   quart
jinja2==3.1.6
    # via
    #   flask
    #   quart
kafka-python==2.2.15
    # via ugc-1-sprint-8 (pyproject.toml)
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   quart
    #   werkzeug
mypy==1.17.0
    # via ugc-1-sprint-8 (pyproject.toml)
mypy-extensions==1.1.0
    # via mypy
opentelemetry-api==1.35.0
    # via
    #   ugc-1-sprint-8 (pyproject.toml)
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-jaeger==1.21.0
    # via ugc-1-sprint-8 (pyproject.toml)
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
    # via opentelemetry-exporter-jaeger
opentelemetry-exporter-jaeger-thrift==1.21.0
    # via opentelemetry-exporter-jaeger
opentelemetry-sdk==1.35.0
    # via
    #   ugc-1-sprint-8 (pyproject.toml)
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
opentelemetry-semantic-conventions==0.56b0
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   aiokafka
    #   pytest
pathspec==0.12.1
    # via mypy
pluggy==1.6.0
    # via pytest
priority==2.0.0
    # via hypercorn
protobuf==4.25.8
    # via googleapis-common-protos
psutil==7.0.0
    # via ugc-1-sprint-8 (pyproject.toml)
pydantic==2.11.7
    # via
    #   ugc-1-sprint-8 (pyproject.toml)
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via ugc-1-sprint-8 (pyproject.toml)
pygments==2.19.2
    # via pytest
pytest==8.4.1
    # via ugc-1-sprint-8 (pyproject.toml)
python-dotenv==1.1.1
    # via pydantic-settings
pytz==2025.2
    # via clickhouse-driver
quart==0.20.0
    # via ugc-1-sprint-8 (pyproject.toml)
redis==6.2.0
    # via ugc-1-sprint-8 (pyproject.toml)
structlog==25.4.0
    # via ugc-1-sprint-8 (pyproject.toml)
taskgroup==0.2.2
    # via hypercorn
tenacity==9.1.2
    # via ugc-1-sprint-8 (pyproject.toml)
thrift==0.22.0
    # via opentelemetry-exporter-jaeger-thrift
tomli==2.2.1
    # via
    #   hypercorn
    #   mypy
    #   pytest
typing-extensions==4.14.1
    # via
    #   aiokafka
    #   exceptiongroup
    #   hypercorn
    #   mypy
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   structlog
    #   taskgroup
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzlocal==5.3.1
    # via clickhouse-driver
werkzeug==3.1.3
    # via
    #   flask
    #   quart
wsproto==1.2.0
    # via hypercorn
zipp==3.23.0
    # via importlib-metadata
